chore: remove unused types and tests
